tests/lib/prepare-restore.sh: clean out snapd changes and snaps before purging
We have a bug where trying to purge the snapd debian package at the same time
that there is an on-going auto-refresh will cause snapd to die and then trigger
snap-failure to start, which re-writes some of the files that the debian
package maintainer scripts deleted, which thus causes the purge to fail because
after the maintainer scripts return, the /var/lib/snapd directory is not empty
and thus isn't removed by apt and the operation fails.

This code sidesteps that issue by cancelling all ongoing changes, then removing
all snaps, then purging snapd. These things should move into the debian
maintainer postrm scripts eventually since this is a real operation that could
happen in the real world, but for now to have more stable spread tests, do this
manually by ourselves.

Also add a bit more debug about the failure condition on the off-chance it does
happen again.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>