boot,gadget,devicestate: create "aux-key" for fde-hooks v2
The new secboot code will needs an additional key called "aux-key"
that is used to bind a model to the disk encryption key. This
commit creates such a key when we use FDE hooks and pases it
around from install to the fde-hook.

It also prepares for TPM to use the aux-key but it is not used
there yet.