boot/flags.go: rename RunModeRootfs to HostUbuntuDataForMode + return list
Return a list of locations, since there could hypothetically be multiple mount
points for ubuntu-data, such as in run mode where there is both "/" and
"/run/mnt/data".

Also verify that the install mode directory actually exists before returning
it. While we expect that this function will primarily be used during the
install-device hook which runs by definition after the partitions are created,
it is possible someday we could use this function at a time in install mode
where the partitions haven't been created and thus the mount point doesn't
exist.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>