bootloader/assets: add support for snapd_full_cmdline_args to grub assets
Add support for snapd_full_cmdline_args variable that carries a set of command
line arguments that shall be used in place of the static and extra ones. Thus
the contents of the new variable effectively become the list of arguments that
directly follows `snapd_recovery_mode=.. snapd_recovery_system=...` in the
kernel command line passed by grub.

The boot config edition is not bumped yet, thus it will not trigger an automatic
boot config update when a new snapd gets installed.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>