o/configstate/configcore/vitality: fix RequireMountedSnapdSnap bug
There currently is a bug where if you apply vitality-hint to a snap service on
UC18+, the service unit will be overwritten without the Requires that prevents
other bugs around starting up very early before snapd is available.

The fix for this is to apply the same logic to configcore vitality handler,
so that the proper set of options is used when re-generating service units that
are included in the vitality-ranking.

Also get rid of an unnecessary loop which was rewriting the service units too
many times.

Finally, add a simple regression spread test about this scenario.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>