systemd, o/servicestate: move InactiveEnter... to systemd package; refactor
Refactor the bug fixing restarting bits of ServiceManager.Ensure() to a
separate function so it can be more easily tested and is simpler to read.

Also use the exported constant for usr-lib-snapd.mount, use the newly exported
systemd.InactiveEnterTimestamp method to get the times for units, and fix some
subtle bugs such as not running for the snapd snap, using snapstate for getting
the device context for easier mocking, and not passing a nil timings to
StartServices which panics on a nil timings.

Also add a MockEnsureSnapServices to easily disable running this logic from
other packages.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>