o/servicestate/servicemgr_test.go: refactor systemctl checks
The issue is that the cleanup function that was deferred was being run as part
of the tear down function, and thus the Check() was not being raised up.
Instead making it something that is deferred directly by the Test*() functions
ensures that the fixtures are still properly setup and so they fail the test
properly.

Also fix 2 unit tests which were passing, but this was not caught due to the
above issue, and would only fail with the tests/unit/go spread test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>