o/servicestate/servicemgr.go: truncate all times to second precision
When we switched to using systemctl show --property InactiveEnterTimestamp
instead of using d-bus, we lost a significant amount of precision of the
timestamp, and now we only get times accurate to the second, while the time
precision for the modification time of the usr-lib-snapd.mount unit file is
much more accurate, ironically actually leading to the seemingly true fact that
the usr-lib-snapd.mount file was modified _after_ usr-lib-snapd.mount was
successfully stopped, since we only have the time that the usr-lib-snapd.mount
unit being stopped to the second. For example, these are the times from a VM:

usr-lib-snapd.mount file modification time = 2021-04-21 02:58:03.142412741
usr-lib-snapd.mount InactiveEnterTimestamp = 2021-04-21 02:58:03
snap.test-snap.svc1.service InactiveEnterTimestamp = 2021-04-21 02:58:03

This means that the time range we end up forming is the nonsensical range:

[ 2021-04-21 02:58:03.142412741 , 2021-04-21 02:58:03 ]

and the code is written such that the candidate timestamp of
2021-04-21 02:58:03 is less than the lower end of the range, and thus must be
outside of the range.

The fix implemented here is to truncate all times to the least precise time we
have, which is that of systemctl, meaning that all times get truncated to
seconds and the above example then becomes:

usr-lib-snapd.mount file modification time = 2021-04-21 02:58:03
usr-lib-snapd.mount InactiveEnterTimestamp = 2021-04-21 02:58:03
snap.test-snap.svc1.service InactiveEnterTimestamp = 2021-04-21 02:58:03

and has the very specific range of a single second:

[ 2021-04-21 02:58:03 , 2021-04-21 02:58:03 ]

which at least does contain the time of the snap service being stopped and thus
ends up fixing the bug.

In addition, in order to have the most predictable behavior here, we should
also always fix our time range here so that if the lower time bound somehow
ends up being in the future significantly compared to the upper time bound, we
just fix the range to be only the upper time bound, since that number at least
should be of the same precision that the candidate snap service stop times.

We should eventually switch back to using D-Bus for getting this property to
have greater precision.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>