overlord/servicestate: implement shortcut for file mod time before boot time
Implement a TODO in the code about taking a shortcut where we know that the
usr-lib-snapd.mount unit file was written before the time of the current boot,
we know that no services could have been killed due to a refresh of snapd
during this boot (and if we rebooted during the change, but after
usr-lib-snapd.mount was written then services would not be killed and we would
revert anyways).

Add some unit tests around this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>