tests/core: test that snapd when refreshing falls back to rebooting for bug
In the case of failing to start services that we know got killed by systemd
when refreshing snapd (i.e. LP #1924805), snapd should reboot the system to try
and get back to a known good state. Test that this is the case and that snapd
is able to recover itself after a reboot to complete the refresh whilst still
rewriting the unit's definitions.

Also, due to the nature of the shutdown command being a symlink to systemctl on
Ubuntu Core, we need to mock the systemctl call but only for shutdown 
operations, passing through the rest of the systemctl commands to the real
systemctl via copying the original systemctl binary somewhere else so it can
still be executed. This code now somewhat duplicates that of the 
uc20-recovery.sh setup script, so we should unify those together some day.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>