quota: simplify AddAllNecessaryGroup implementation
It is simpler to just walk up to the top of the tree and then get the full
sub-tree below the root group. This is not an efficient implementation for
large trees, but we do not expect there to be large trees of groups so this is
fine for now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>