quota: new package for managing resource groups
* quota: add new quota package

With some helpers to get related quotas from a single quota reference and to
initialize the references from a deserialized set of quotas.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: prefix slice names with snap.<group-name>

This ensures that we avoid the default systemd namespace for slices to make
name squatting of the form "user1000.slice" ineffective as it would be named
"snap.user1000.slice" in reality.

Thanks to Maciej for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: use more careful arithmetic to avoid max int overflow

Thanks to Maciej for spotting this

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: adjust doc-comment for MemoryLimit

Thanks to Maciej for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: make quota group names at least 2 characters long

This matches other name requirements we have in snapd like the name of snaps.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: simplify AddAllNecessaryGroup implementation

It is simpler to just walk up to the top of the tree and then get the full
sub-tree below the root group. This is not an efficient implementation for
large trees, but we do not expect there to be large trees of groups so this is
fine for now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: disallow capitalized quota names

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: rename CompleteCrossReferences to ResolveCross...

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: disallow empty group names with specific error

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: check that the sub-group is not the same as the parent earlier

This gets a specific error message, so we should check it earlier, before
validating the desired sub-group.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: validate that groups do not have circular references

This won't be possible with NewGroup and NewSubGroup, but it will be possible
with corrupted or invalid serialization of all the groups when we use
ResolveCrossReferences on a set of groups.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: expand cross-checks in ResolveCrossReferences

Specifically, ensure that sub-groups reference their parents and vice versa.

Also add unit test cases for ResolveCrossReferences.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: rename subTree() -> tree() + include the group itself in the output

Also add full unit tests, and sort the output from AllQuotaGroups.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: add doc-comments for QuotaGroupSet and AllQuotaGroups

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/naming, quota: move quota grp validation to naming pkg

Also adjust some unit tests, but this means that now we will support dashes in
group names, but we have yet to actually escape the dashes in group names.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* quota: properly escape dashes in group names for slice file names

We need to escape the "-" character in group names, since this is used by
systemd to identify "/" in the path. Add some unit tests for this case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/naming/validate.go: add tests for quota groups, chg err message

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/naming: adjust ValidateQuotaGroup and its tests

test was testing ValidSnap for the happy cases.

use "quota group" for consistency in errors

hide the regexp as is the case for snap names

* quota: have a packate doc comment.

* quota: clarify Group.Name doc comment

* quota: test SliceFileName more

* quota: move NewGroup and give it a doc comment.

* quota: remove code that is not necessary

* quota: simplify the operating assumptions for QuotaGroupSet

* quota: add more ResolveCrossReferences tests

fix what seems like a bug in validate memory limit check

* quota: add TODO?

* snap/quota: move quota package here

TODO about another package placement

* snap/quota: rm unnecessary type, fix doc-comments, fix test name

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

Co-authored-by: Samuele Pedroni <pedronis@lucediurna.net>