o/servicestate/quotas: UpdateQuotas -> PatchQuotasState; return new group state
It is convenient when updating quotas to get back the new full state of quotas
since many times we will have the pattern of calling UpdateQuotas to set the
state, then immediately calling SnapServiceOptions many times to calculate
input to provide to EnsureSnapServices, and thus being able to get the updated
state back from UpdateQuotas to provide to SnapServiceOptions reduces the calls
to get quotas from state, and thus also reduces the need to re-verify what's in
state every time the way that AllQuotas currently does.

Also, since we will need another similarly named function, UpdateQuota which 
will be used by daemon, so for simplicity's sake rename this function to reduce
confusion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>