wrappers/services_test.go: addl case for only snap with sub-group, parent gen'd
Add a unit test for the case where we have a quota group that only has a
sub-group in it, no snaps, and the generated snap service is in the sub-group.
The code in this case should generate both the parent quota group and the
sub-group systemd slice units automatically.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>