o/servicestae/servicemgr_test.go: increment systemctlCalls when not handled
This had a bug previously where calls past the expected number did not cause
the test to fail if the systemctl command's output and error status was not
checked.

Now we will fail properly at the end with the function that ensures the number
of systemctl calls matches what we expected.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>