o/servicestate/quota_control.go: restart slices + svcs after group modification
When we add or remove a snap from a slice (or modify the slice itself), we need
to restart the slice unit and also restart the snaps such that systemd moves
the processes over, as it does not move them over just with systemctl
daemon-reload.

Also update the tests to check that systemctl is used as expected.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>