o/servicestate_quota_control: detect inconsistent parent in RemoveQuota
This ensures that we assign SubGroups of the parent as nil directly if there
are no other children (before the code was assigning an empty, non-nil list).

It also ensures that if the parent group did not contain a reference to the
child that we are trying to remove (which is an inconsistent state), then we
will fail to do anything whereas previously we would gloss over this fact and
nothing would be raised to the user.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>