o/servicestate/quota_control.go: remove unnecessary error checking
We can never hit these errors as the code stands, since these errors only show
up when we have inconsistent state, which we cannot have since AllQuotas would
have errored out at the top already. As such just remove the error handling
code and we can take a look at fixing it up later to make it more robust as we
might someday need to be able to delete groups that are making the state
inconsistent.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>