o/servicestate, many: check systemd version before using quota groups
We need systemd version 205+ in order to properly use slice units, so check
this for each function.

Note: we could cache the result of this check since systemd versions shouldn't
change during a boot, but the check should be relatively cheap to perform and
is simpler to handle here by just calling it every time (if a bit verbose in
tests).

Also update the spread test to skip trusty since it has systemd 204.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>