o/servicestate: refactor systemd version check
We now only call the checking function once at package initialization which is
simpler and requires less global state, but it does mean we need to separately
export and then call the checking function from the unit test which uses the
"real" systemctl command in the code to check the version.

Thanks to Samuele for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>