o/servicestate/quota_control.go: refactor systemctl + checks for slice mods
When modifying slice units, we have a few different cases that need to be
handled which require systemctl operations.

1) Snaps are added to a slice (also another case for removal in a followup PR)
2) A group is newly created
3) A group is removed
4a) A group has it's memory limit increased
4a) A group has it's memory limit decreased

For 1, we don't need to modify the slice at all, but the services need to be
restarted.

For 2, the slice unit needs to be started, and any new snaps in the group are
subject to 1.

For 3, the slice unit has to be stopped, and any snaps that were in the group
are subject to 1

For 4a, we just need to call systemctl daemon-reload

For 4b, we need to first stop any services currently in the slice, then call
systemctl daemon-reload, then start any stopped services. This is because the
systemctl daemon-reload call could fail if there are currently processes in the
cgroup which are using more memory than the new memory limit (since it is being
decreased), and it fails in an unfortunate way where the systemctl call does
not fail, but the old higher limit remains in effect but systemd reports the
new, not currently enforced limit, as being in effect.

Becuase 4b is complicated to handle, we exclude it by preventing any
modification to an existing quota group that would decrease the memory limit.
We may someday wish to enable this use-case but it will take more code to
properly implement.

Also refactor the restarting of services in ensureSnapServicesForGroup slightly
to release the state lock while services are being started, and restart them
in their startup order rather than stopping them all and then starting them up
one by one.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>