overlord/servicestate: remember if acting on the entire snap
If only the snap name is given on the command line, do not always treat
it as if all services were mentioned, but rather pass the information
down to the command executors: their behavior might change depending on
whether the user specified a list of services or just the snap name.

This commit does not change the behavior of snapd.