release-tools/changelog.py: refactor regexp + file reading/writing
Use a more general regular expression with a capturing group to enable
modifying just the version number we need to change without needing the hackish
manual whitespace bits inserted as part of the version number to replace. This
also enables us to confirm that only one version modification was made to the
file whereas previously we could be accidentally modifying more than one line
without realizing it.

Also optimizing the reading and writing of files so we only read each file we
need once, modify it in a memory buffer, then write it back out once. This
should make the script faster and less likely to fail with half modified files
where only the version number was changed but not the changelog itself for
example with the Fedora spec file.

Thanks to Maciej for all the suggestions!

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>