interfaces/apparmor: limit the number of jobs when running with a single CPU
When the system has a single CPU, snapd would not pass any jobs flags to
apparmor_parser. In such case, the defaults would be used, which allow for up to
8 jobs to be started in parallel. What means, that with the way we pass the
flags, in a system with a single CPU would run the same number of jobs as 10 CPU
one (as we use #CPU - 2 there). Since the system has just 1 CPU, avoid the fork
overhead and use the main process for compilation.

See [1] for reference.

1. https://gitlab.com/apparmor/apparmor/-/blob/7dcf013bcab9548582734db244ba74f09449f9c1/parser/parser_main.c

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>