overlord, overlord/devicestate: allow for reloading modeenv in devicemgr when testing
When running managers tests, the device manager is created as part of the test
setup. The modeenv is only read once, when the manager is created, thus making
it problematic for tests that modify the modeenv to observe the desired changes.
Add a helper that allows for reloading the modeenv in tests.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>