release-tools/changelog.py: specify the LP bug for the release as an arg too
The LP bug for the release is different for every new major release, so we need
to specify this when generating the changelog.

We could eventually figure out what it should be for minor releases, but for
now just make it mandatory to specify it in the arguments to the script.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>