tests/main/snapd-snap: don't use $PROJECT_PATH as a dir for tests.cleanup
This breaks an invariant check, instead continue to use the test's normal $PWD,
but use a manually written script to execute snapcraft clean in $PROJECT_PATH
because trying to do `cd $PROJECT_PATH && snapcraft clean` does not work with
tests.cleanup.

We also don't need to manually delete the built snap if we just move it into
the test's $PWD.

Also add --purge to removing the snaps.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>