release-tools/changelog.py: add function to generate github release template
This just basically adds a basic templated header to the changelog entries and
makes the changelog entries align with proper markdown lists etc.

Eventually it would be super nice to be able to use the GitHub API to actually
create the release after it has been tagged and pushed, etc. with this included
automatically, but that will require an API key and also code, so it is left as
a future improvement.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>