release-tools/changelog.py: also fix opensuse changelog date format
Thanks to Maciej for sorting out the right format here, as we were using an
incorrect or inconsistent format for the past few releases, likely cargo
culting the first incorrect format that was introduced last year.

Also we should just always use UTC for this, so we can simplify the code
slightly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>