systemd: improve systemctl error reporting
We recently had a build failure on trusty in a schroot environment.
It turned out that the issue was that no systemctl binary was installed
there. But the error was a bit misleading:
```
[--version] failed with exit status 0:
```

This commit improves the error reporting for non-exit code errors
and errors that do not have an associated exit status.