snap/quota: adjust workaround for erroneous systemd reported usage value
This is a more specialized handling of the erroneous value reported from
systemd, since we actually do have cases where there can be non-zero memory
usage reported for otherwise empty (no tasks) cgroups. As such, only fix the
memory usage to 0 when there are both no tasks, and when the memory value
reported is the erroneous value of 16 exbibytes.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>