o/servicestate/quota_control.go: use handlers implementation (temporarily)
These exported methods will now temporarily call the underlying handlers
directly. This is done to demonstrate that the handlers are working properly
with all the existing unit tests.

The next patch here will switch all the tests for the exported methods over to
testing the handlers directly instead of via the exported methods so we can
switch the exported methods over to returning tasksets which are not executed
directly and instead are put into a change which the overlord will execute.

Also delete the unnecessary unit tests for the happy path of the handlers that
duplicated the unit tests in quota_control_test.go.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>