o/servicestate/quota_control_test.go: test the handlers directly
These unit tests exercise the more complicated logic which now lives in the
handlers, so call the handlers directly.

Also add some basic happy path tests for the exported functions to be sure that
they are still working, but the exported functions will shortly be refactored
to return task sets and not perform the logic directly.

Also adjust the error message from the handler implementation to match the
original implementation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>