overlord/servicestate: improve test robustness with locking
Tests were not undoing the last Lock() operation on the state. We now
issue a "defer Unlock()" right after the first Lock() invocation to
ensure that the mutex will get unlocked when the test function stops.

Special care must be taken with check assertions occurring after the
state has been explicitly Unlocked() again, because a test failure would
cause the test function to terminate and the deferred operation would
attempt to unlock an unlocked mutex. That's why we avoid assertions
while in unlocked state, and instead just capture the subject of our
checks into variables and do the assertion only when the state has been
locked again.