o/servicestate/quota_control.go: bump minimum systemd version up to 230
Xenial and older systemds suffer from a related bug to the one that we fixed in
snapcore/snapd#10354, but this variant of the bug is worse in that it triggers
systemd to report other groups that have real usage as being the maximum value
for a 64-bit integer and also makes our workaround ineffective since the tasks
in a slice gets reported as being the maximum 64-bit integer value as well.

Also add a spread test to confirm that other systems are not effected.

The work-around we added is no longer needed so it will be rolled back in a
future commit.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>