sandbox/cgroup: skip own cgroup when freezing
With v2 hierarchy, we use the tracking cgroup for identifying processes of a
snap. When the snap application runs, the snap run -> snap-confine ->
snap-update-ns chain is placed in the tracking scope early on. Which means that
when the freezing code is executed, the process may already part of the group.
In such event, we need to skip our own cgroup, so that we do not freeze
ourselves.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>