o/servicestate: refactor meter handling, eliminate some common parameters
We can get rid of the meter argument to most of the helpers if we delay
creating the meter until when we need it doing systemd operations. When we do
those operations however, we also need to be careful to use the right meter,
since some operations are done with the state lock and some are done without
the state lock.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>