snap: fix TestInstallNoPATH unit test failure when SUDO_UID is set
The TestInstallNoPATH unit test is currently failing when the SUDO_UID
environment is set. This happens e.g. in a pbuilder environment.

To reproduce:
```
$ SUDO_UID=100 go test -check.f TestInstallNoPATH

FAIL: cmd_snap_op_test.go:229: SnapOpSuite.TestInstallNoPATH

cmd_snap_op_test.go:252:
    c.Check(s.Stderr(), testutil.MatchesWrapped, `Warning: \S+/bin was not found in your \$PATH.*`)
... wrapped string = ""
... expected string = "Warning: \\S+/bin was not found in your \\$PATH.*"
```

The fix is to ensure that SUDO_UID is unset in this test.