HACKING.md: update instructions for go1.16+
With go1.16 we need to disable go modules or our instructions
will fail at the "go get" stage because go will try to init
using go modules.

This can be dropped once we move to go 1.13 but right now
out instructions do not work for people on Ubuntu 21.04 or
similar new distro releases.