interfaces/apparmor: allow snap-update-ns to access /sys/devices/system/cpu/online
Glibc 2.27 may try to access this file when creating a new arena for malloc.
This happens when glibc calls __get_nprocs() behind the scenes, and
/sys/devices/systen/cpu/online is the first of the files on which open() is
attempted.

Putting this in Go context, with details outlined in
https://github.com/golang/go/issues/25628, the problem was only seed on i386
systems when binaries are built using cgo. It just so happens that
snap-update-ns needs to be built with cgo and linked statically with glibc.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>