Add --ask parameter and documents portal support to snapctl user-open
Where possible, snapctl user-open and by extension /bin/xdg-open
will attempt to use the documents portal, such that if a file
is writable by the snap sending the request, it is writable by the
application receiving the request.
This requires xdg-desktop-portal >= 1.8.1 to work snap to snap
but should work with earlier versions snap to flatpak.

The --ask flag enables the snap to request that the portals UI
is shown unconditionally, ignoring for example previously learnt
associations.